feat(kubernetes): Move com.netflix.spinnaker.clouddriver.orchestration.OperationDescription to kork-moniker
Moved the OperationDescription from clouddriver-api to kork-moniker and updated all the relevant references.
This will allow us to pass Manifest Description to the Namer class.
