--- clouddriver-api/src/main/java/com/netflix/spinnaker/clouddriver/orchestration/OperationDescription.java
@@ -1,4 +1,0 @@
-package com.netflix.spinnaker.clouddriver.orchestration;
-
-/** Marker interface for inputs to an {@link AtomicOperation}. */
-public interface OperationDescription {}
