+++ clouddriver-core/src/main/java/com/netflix/spinnaker/clouddriver/deploy/DeployDescription.java
@@ -16,8 +16,8 @@
 
 package com.netflix.spinnaker.clouddriver.deploy;
 
-import com.netflix.spinnaker.clouddriver.orchestration.OperationDescription;
 import com.netflix.spinnaker.clouddriver.orchestration.events.OperationEvent;
+import com.netflix.spinnaker.orchestration.OperationDescription;
 import java.util.Collection;
 import java.util.Collections;
