+++ clouddriver-core/src/main/java/com/netflix/spinnaker/clouddriver/orchestration/OperationsService.java
@@ -32,6 +32,7 @@ import com.netflix.spinnaker.clouddriver.security.AccountCredentialsRepository;
 import com.netflix.spinnaker.clouddriver.security.AllowedAccountsValidator;
 import com.netflix.spinnaker.kork.exceptions.SystemException;
 import com.netflix.spinnaker.kork.web.exceptions.ExceptionMessageDecorator;
+import com.netflix.spinnaker.orchestration.OperationDescription;
 import com.netflix.spinnaker.security.AuthenticatedRequest;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
