+++ clouddriver-elasticsearch/clouddriver-elasticsearch.gradle
@@ -6,6 +6,7 @@ dependencies {
   implementation "com.netflix.frigga:frigga"
   implementation "io.searchbox:jest:6.3.1"
   implementation "io.spinnaker.kork:kork-exceptions"
+  implementation "io.spinnaker.kork:kork-moniker"
   implementation "io.spinnaker.kork:kork-retrofit"
   implementation "io.spinnaker.kork:kork-security"
   implementation "com.squareup.retrofit:retrofit"
