+++ clouddriver-security/clouddriver-security.gradle
@@ -9,6 +9,7 @@ dependencies {
   implementation "io.spinnaker.fiat:fiat-api:$fiatVersion"
   implementation "io.spinnaker.fiat:fiat-core:$fiatVersion"
   implementation "io.spinnaker.kork:kork-core"
+  implementation "io.spinnaker.kork:kork-moniker"
   implementation "org.codehaus.groovy:groovy"
   implementation "org.slf4j:jcl-over-slf4j"
   implementation "org.springframework.boot:spring-boot-starter-web"
